+++ NEWS.md
@@ -1,6 +1,7 @@
 # dsem 2.1.0
 
 * Change to requiring `family` options that agree with `glm`
+* Adding `family = gaussian_fixed_sd()` option to pre-specify a known standard deviation
 
 # dsem 2.0.0